added cofusion matrix to the GUI and created one run script main.py and finalized the project

+++ classification.py
@@ -28,49 +28,56 @@ constant_random_state = 42
 
 def apply_linear_regression(x_train,y_train,x_test,y_test,file_name) :
 
-    logreg = LogisticRegression(max_iter = 3000 ,solver='lbfgs')
-
-    logreg.fit(x_train,y_train)
-    print(logreg.coef_)
-    y_pred=logreg.predict(x_test)
-    logregCf = confusion_matrix(y_test, y_pred)
-
     fn = "MemoizedTrainResults/LR_" + file_name
     fn = fn.replace("csv","joblib")
-    joblib.dump(logreg, fn)
-    
-    
+    logreg = -1
+
+    if os.path.exists(fn) :
+        logreg = joblib.load(fn)
+    else :
+        logreg = LogisticRegression(max_iter = 3000 ,solver='lbfgs')
+        logreg.fit(x_train,y_train)
+        joblib.dump(logreg, fn)
+
+    y_pred=logreg.predict(x_test)
+    logregCf = confusion_matrix(y_test, y_pred)
 
     return [accuracy_score(y_test,y_pred),logregCf]
 
 def apply_discession_tree(x_train,y_train,x_test,y_test,file_name) :
-    dec_tree = DecisionTreeClassifier(
-        random_state = constant_random_state ,max_features = int(math.log2(int(64/2))))
-
-    dec_tree.fit(x_train,y_train)
-    y_pred = dec_tree.predict(x_test)
-    decTreeCf = confusion_matrix(y_test, y_pred)
 
     fn = "MemoizedTrainResults/DT_" + file_name
     fn = fn.replace("csv","joblib")
+    dec_tree = -1
 
-    joblib.dump(dec_tree, fn)
+    if (os.path.exists(fn)) :
+        dec_tree = joblib.load(fn)
+    else :
+        dec_tree = DecisionTreeClassifier(random_state = constant_random_state ,max_features = int(math.log2(int(64/2))))
+        dec_tree.fit(x_train,y_train)
+        joblib.dump(dec_tree, fn)
 
-    
+    y_pred = dec_tree.predict(x_test)
+    decTreeCf = confusion_matrix(y_test, y_pred)
+
     return [accuracy_score(y_test,y_pred),decTreeCf]
 
 def apply_svm_linear_kernel(x_train,y_train,x_test,y_test,file_name) :
 
-    clf = svm.SVC()
-    clf.fit(x_train, y_train)
-    y_pred = clf.predict(x_test)
-    svmCf= confusion_matrix(y_test, y_pred)
-
     fn = "MemoizedTrainResults/SVM_" + file_name
     fn = fn.replace("csv","joblib")
-    joblib.dump(clf, fn)
-    
+    clf = -1
 
+    if (os.path.exists(fn)) :
+        clf = joblib.load(fn)
+    else :
+        clf = svm.SVC()
+        clf.fit(x_train,y_train)
+        joblib.dump(clf, fn)
+
+    y_pred = clf.predict(x_test)
+    svmCf= confusion_matrix(y_test, y_pred)
+ 
     return [(metrics.accuracy_score(y_test, y_pred)),svmCf]
 
 
@@ -84,6 +91,9 @@ mx_discession_tree_after_normalization = 0.0
 
 mx_svm_before_normalization = 0.0
 mx_svm_after_normalization = 0.0
+
+CMB = []
+CMA = []
 
 i = 1
 for file_name in cleaned_data_list :
@@ -103,14 +113,12 @@ for file_name in cleaned_data_list :
 
     ros = RandomOverSampler(random_state=0)
     X_train_ros, y_train_ros= ros.fit_resample(x_train, y_train)
-    # print(sorted(Counter(y_train_ros).items()))
 
     rf = RandomForestClassifier()
     ros_model = rf.fit(X_train_ros, y_train_ros)
     ros_prediction = ros_model.predict(x_test)
 
     smote = SMOTE(random_state=constant_random_state)
-    # print(classification_report(y_test, ros_prediction))
     X_train_smote, Y_train_smote= smote.fit_resample(x_train, y_train)
 
     
@@ -118,9 +126,15 @@ for file_name in cleaned_data_list :
     x_train = X_train_smote
     y_train = Y_train_smote
 
-    linear_regression_accuracy = apply_linear_regression(x_train,y_train,x_test,y_test,file_name)[0]
-    discession_tree_accuracy = apply_discession_tree(x_train,y_train,x_test,y_test,file_name)[0]
-    svm_linear_kernel_accuracy = apply_svm_linear_kernel(x_train,y_train,x_test,y_test,file_name)[0]
+    lin_res = apply_linear_regression(x_train,y_train,x_test,y_test,file_name)
+    dt_res = apply_discession_tree(x_train,y_train,x_test,y_test,file_name)
+    svm_res = apply_svm_linear_kernel(x_train,y_train,x_test,y_test,file_name)
+
+    linear_regression_accuracy = lin_res[0]
+    discession_tree_accuracy = dt_res[0]
+    svm_linear_kernel_accuracy = svm_res[0]
+
+
 
     print(file_name)
     print("logistic regression : ",linear_regression_accuracy)
@@ -132,10 +146,20 @@ for file_name in cleaned_data_list :
         mx_linear_before_normalization = linear_regression_accuracy
         mx_discession_tree_before_normalization = discession_tree_accuracy
         mx_svm_before_normalization = svm_linear_kernel_accuracy
+
+        CMB.append(lin_res[1])
+        CMB.append(dt_res[1])
+        CMB.append(svm_res[1])
+
     else :
         mx_linear_after_normalization = linear_regression_accuracy
         mx_discession_tree_after_normalization = discession_tree_accuracy
         mx_svm_after_normalization = svm_linear_kernel_accuracy
+
+        CMA.append(lin_res[1])
+        CMA.append(dt_res[1])
+        CMA.append(svm_res[1])
+        CMA = np.array(CMA).tolist()
 
     
     i += 1
@@ -151,12 +175,13 @@ print("max svm before normalization : ",mx_svm_before_normalization)
 print("max svm after normalization : ",mx_svm_after_normalization)
 
 
+print(CMA)
 
-results = str(mx_linear_before_normalization) + ','+str(mx_linear_after_normalization) + '\n'
-results += str(mx_discession_tree_before_normalization) +','+ str(mx_discession_tree_after_normalization) + '\n'
-results += str(mx_svm_before_normalization) + ','+str(mx_svm_after_normalization) + '\n'
+
+results = str(mx_linear_before_normalization) + ','+str(mx_linear_after_normalization) + "," + str(CMA[0][0][0]) + "," + str(CMA[0][0][1]) + ","+ str(CMA[0][1][0]) + "," + str(CMA[0][1][1]) + '\n'
+results += str(mx_discession_tree_before_normalization) +','+ str(mx_discession_tree_after_normalization)+ "," + str(CMA[1][0][0]) + "," + str(CMA[1][0][1]) + ","+ str(CMA[1][1][0]) + "," + str(CMA[1][1][1]) + '\n'
+results += str(mx_svm_before_normalization) + ','+str(mx_svm_after_normalization)+ ',' + str(CMA[2][0][0]) + "," + str(CMA[2][0][1]) + ","+ str(CMA[2][1][0]) + "," + str(CMA[2][1][1]) + '\n'
+
 file = open("results.txt",'w')
 file.write(results)
 file.close()
-
-input()
